+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # CHANGELOG
 
+### 3Jan2019
+- Added `scale` property to `OCT` class and modified `analysis\alignImages.m` accordingly
+
 ### 2Jan2019
 - Wrote `HistogramSlider.m` to compare 1D histograms with segmented RPE, ILM and choroid.
 - Wrote `HistogramPeakSlider.m` to cooperate with `ChoroidApp.m` for interactive identification of choroid boundary control points using the 1D histograms.